Add TVDB series/episode translation support
Adds the ability to explicitly specify a translation language for TVDB,
and automatically handle translations of series and episode titles into
that language. This can ensure that file names are properly translated
to match your expectations (e.g. turning "ポケットモンスター" into
"Pokémon").

This behaviour is triggered by a new, optional key under the TVDB API
configuration called "language", which if present and set to a non-empty
value, will be used as the translation language. Two additional keys
under this section, which are not optional if the language is set,
provide the translation API endpoints for the series and episode names,
respectively.